fix: bug when saving multiple times on the same namespace and name

+++ storage/sqlite/migrations/000003_create_snapshots_protobuf_files_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS snapshots_protobuf_files(
-	snapshot_id BIGINT,
-	file_id BIGINT,
+	snapshot_id INTEGER,
+	file_id INTEGER,
 	CONSTRAINT fk_snapshot FOREIGN KEY(snapshot_id) REFERENCES snapshots(id),
 	CONSTRAINT fk_file FOREIGN KEY(file_id) REFERENCES protobuf_files(id)
 );
